+++ postcss.config.js
@@ -2,5 +2,6 @@ module.exports = {
   plugins: {
     autoprefixer: {},
     'postcss-flexbugs-fixes': {},
+    'css-mqpacker': {},
   },
 };