+++ src/utils/logger/LogModel.js
@@ -1,205 +1,276 @@
 /* eslint-disable no-console */
-import { isString, isPlainObject } from '@/utils/inspect';
+import {
+  isNil,
+  isInteger,
+  isArray,
+  isPlainObject,
+  isEmptyObject,
+} from '../inspect';
+import { cloneDeep } from '../index';
 import { STYLES } from './constants';
 
-/**
- * @example
- * logger.log(new LogModel({
- *   heading: 'methodName()',
- *   message: 'Data is invalid!',
- *   data: {
- *     name: nameValue,
- *     date: dateValue,
- *     props,
- *   },
- * }));
- */
 export default class LogModel {
   /**
    * @param {Object} config
-   * @param {String} config.heading Заголовок после тега (например имя функции)
-   * @param {String} config.message Описание (вторая строка лога)
-   * @param {Object} config.data Произвольные данные
+   * @param {Array} [config.tags] Список тегов после тега области (первая строка лога)
+   * @param {Array} [config.messages] Список строк описания (вторая и далее строки лога)
+   * @param {Object} [config.data] Именованные свойства
+   * @param {*} [config.epilog] Произвольные данные (после всех вышеописанных данных)
    */
-  constructor(config = {}) {
-    if (config && !isPlainObject(config)) {
-      console.error(
-        '[LogModel:constructor]:\n',
-        'The "config" property is invalid!\n',
-        '| Given value:', config,
-      );
-      return;
+  constructor(config) {
+    this.props = {};
+    this.template = '';
+    this.substrings = [];
+
+    const isValid = LogModel.validateConfig(config);
+
+    if (isValid) {
+      this.props = cloneDeep(config);
+    }
+  }
+
+  /**
+   * @param {Object} config
+   * @returns {Boolean}
+   * @private
+   */
+  static validateConfig(config) {
+    const error = (...args) => {
+      console.error('[LogModel] validateConfig()\n', ...args);
+    };
+
+    if (!isPlainObject(config)) {
+      error('The "config" property is invalid!\n', '| Given value:', config);
+      return false;
     }
 
-    const { heading, message, data } = config;
+    const { tags, messages, data } = config;
 
-    if (heading && !isString(heading)) {
-      console.error(
-        '[Logger:constructor]:',
-        'The "heading" option is invalid!\n',
-        '| Given value:', heading,
-      );
-      return;
+    if (tags && !isArray(tags)) {
+      error('The "tags" option is invalid!\n', '| Given value:', tags);
+      return false;
     }
-    if (message && !isString(message)) {
-      console.error(
-        '[Logger:constructor]:',
-        'The "message" option is invalid!\n',
-        '| Given value:', message,
-      );
-      return;
+    if (messages && !isArray(messages)) {
+      error('The "messages" option is invalid!\n', '| Given value:', messages);
+      return false;
     }
     if (data && !isPlainObject(data)) {
-      console.error(
-        '[Logger:constructor]:',
-        'The "data" option is invalid!\n',
-        '| Given value:', data,
-      );
+      error('The "data" option is invalid!\n', '| Given value:', data);
+      return false;
+    }
+
+    return true;
+  }
+
+  /**
+   * Инъекция данных инстанса логгера
+   * @params {*} context
+   * @params {Array} tags
+   * @returns {Boolean}
+   */
+  injectLoggerData(context, tags) {
+    if (!context?.isLoggerInstance) return false;
+
+    this.props.tags = this.props.tags ? [...tags, ...this.props.tags] : tags;
+
+    return true;
+  }
+
+  /**
+   * Подготовить переданные данные
+   * @param {String} level
+   * @private
+   */
+  prepareData(level) {
+    if (isEmptyObject(this.props)) return;
+
+    const {
+      tags, messages, data, epilog,
+    } = this.props;
+
+    this.template = '';
+    this.substrings = [];
+
+    if (tags) {
+      tags.forEach((tag) => {
+        this.appendTag(level, tag);
+      });
+    }
+
+    if (messages) {
+      messages.forEach((message) => {
+        this.appendNewLine();
+        this.appendStartLine(level);
+        this.appendText(message);
+      });
+    }
+
+    if (data) {
+      Object.entries(data).forEach(([key, value]) => {
+        this.appendNewLine();
+        this.appendStartLine(level, '> ');
+        this.appendProperty({ level, key, value });
+      });
+    }
+
+    if (epilog) {
+      this.appendNewLine();
+      this.appendText(epilog);
+    }
+  }
+
+  /**
+   * Добавить свойство объекта с ключом и значением по типу
+   * @param {Object} payload
+   * @param {String} payload.level
+   * @param {String} payload.key
+   * @param {*} payload.value
+   * @private
+   */
+  appendProperty({ level, key, value }) {
+    if (key) {
+      this.appendText(key);
+      this.template += ' %c=>%c ';
+      this.substrings.push(STYLES.TEXT[level]);
+      this.substrings.push('');
+    }
+
+    if (isNil(value)) {
+      this.appendBoolean(value);
       return;
     }
 
-    this.config = config;
-  }
-
-  /**
-   * Вставить сброс строки
-   * @private
-   */
-  setNewLine() {
-    this.template += this.template ? '\n' : '';
-  }
-
-  /**
-   * Вставить цветную вертикальную линию
-   * @param {String} level
-   * @private
-   */
-  setStartLine(level) {
-    this.template += '%c| %c';
-    this.substrings.push(STYLES.TEXT[level]);
-    this.substrings.push(null);
-  }
-
-  /**
-   * Вставить текст цвета текущей темы браузера
-   * @param {*} value
-   * @private
-   */
-  setText(value) {
-    this.template += '%s';
-    this.substrings.push(value);
-  }
-
-  /**
-   * Вставить булево значение выделенное цветом
-   * @param {Boolean | null} value
-   * @private
-   */
-  setBoolean(value) {
-    this.template += '%c%s%c';
-    this.substrings.push(STYLES.TEXT.ORANGE);
-    this.substrings.push(value);
-    this.substrings.push(null);
-  }
-
-  /**
-   * Вставить число выделенное цветом
-   * @param {Number} value
-   * @private
-   */
-  setNumber(value) {
-    this.template += '%c%f%c';
-    this.substrings.push(STYLES.TEXT.BLUE);
-    this.substrings.push(value);
-    this.substrings.push(null);
-  }
-
-  /**
-   * Вставить строку выделенное цветом
-   * @param {String} value
-   * @private
-   */
-  setString(value) {
-    this.template += '%c%s%c';
-    this.substrings.push(STYLES.TEXT.GREEN);
-    this.substrings.push(value);
-    this.substrings.push(null);
-  }
-
-  /**
-   * Вставить объект или массив
-   * @param {Object | Array} value
-   * @private
-   */
-  setObject(value) {
-    this.template += '%O';
-    this.substrings.push(value);
-  }
-
-  /**
-   * Вставить свойство объекта с ключом и значением по типу
-   * @param {String} level
-   * @param {String} key
-   * @param {*} value
-   * @private
-   */
-  setProperty(level, key, value) {
-    this.setText(key);
-    this.template += ' %c=>%c ';
-    this.substrings.push(STYLES.TEXT[level]);
-    this.substrings.push(null);
-
-    if (value === null || value === undefined) {
-      this.setBoolean(value);
+    if (value instanceof Error) {
+      this.appendText(value);
       return;
     }
 
     switch (typeof value) {
       case 'boolean':
-        this.setBoolean(value);
+        this.appendBoolean(value);
         break;
       case 'number':
-        this.setNumber(value);
+        this.appendNumber(value);
         break;
       case 'string':
-        this.setString(value);
+        this.appendString(value);
+        break;
+      case 'symbol':
+        this.appendSymbol(value);
         break;
       case 'object':
-        this.setObject(value);
+        this.appendObject(value);
         break;
       default:
-        this.setText(value);
+        this.appendText(value);
     }
   }
 
   /**
-   * Форматировать переданные данные
-   * @param {String} level
+   * Добавить сброс строки
    * @private
    */
-  format(level) {
-    const { heading, message, data } = this.config;
+  appendNewLine() {
+    this.template += '\n';
+  }
 
-    this.template = '';
-    this.substrings = [];
+  /**
+   * Добавить цветную вертикальную линию
+   * @param {String} level
+   * @param {String} [suffix]
+   * @private
+   */
+  appendStartLine(level, suffix = '') {
+    this.template += `%c| ${suffix}%c`;
+    this.substrings.push(STYLES.TEXT[level]);
+    this.substrings.push('');
+  }
 
-    if (heading) {
-      this.setText(heading);
+  /**
+   * Добавить тэг
+   * @param {String} level
+   * @param {*} value
+   * @private
+   */
+  appendTag(level, value) {
+    if (this.template) {
+      this.template += '%c>';
+      this.substrings.push(STYLES.TAG_SEPARATOR[level]);
     }
+    this.template += '%c%s%c';
+    this.substrings.push(STYLES.TAG[level]);
+    this.substrings.push(value);
+    this.substrings.push('');
+  }
 
-    if (message) {
-      this.setNewLine();
-      this.setStartLine(level);
-      this.setText(message);
-    }
+  /**
+   * Добавить текст цвета текущей темы браузера
+   * @param {*} value
+   * @private
+   */
+  appendText(value) {
+    this.template += '%s';
+    this.substrings.push(value);
+  }
 
-    if (data) {
-      Object.entries(data).forEach(([key, value]) => {
-        this.setNewLine();
-        this.setStartLine(level);
-        this.setProperty(level, key, value);
-      });
-    }
+  /**
+   * Добавить булево значение выделенное цветом
+   * @param {Boolean | null | undefined} value
+   * @private
+   */
+  appendBoolean(value) {
+    this.template += '%c%s%c';
+    this.substrings.push(STYLES.TEXT.BOOLEAN);
+    this.substrings.push(value);
+    this.substrings.push('');
+  }
+
+  /**
+   * Добавить число выделенное цветом
+   * @param {Number} value
+   * @private
+   */
+  appendNumber(value) {
+    const substitution = isInteger(value) ? '%i' : '%f';
+
+    this.template += `%c${substitution}%c`;
+    this.substrings.push(STYLES.TEXT.NUMBER);
+    this.substrings.push(value);
+    this.substrings.push('');
+  }
+
+  /**
+   * Добавить строку выделенное цветом
+   * @param {String} value
+   * @private
+   */
+  appendString(value) {
+    this.template += '%c%s%c';
+    this.substrings.push(STYLES.TEXT.STRING);
+    this.substrings.push(value);
+    this.substrings.push('');
+  }
+
+  /**
+   * Добавить символ
+   * @param {Symbol} value
+   * @private
+   */
+  appendSymbol(value) {
+    this.template += '%c%s%c';
+    this.substrings.push(STYLES.TEXT.SYMBOL);
+    this.substrings.push(value.toString());
+    this.substrings.push('');
+  }
+
+  /**
+   * Добавить объект или массив
+   * @param {Object | Array} value
+   * @private
+   */
+  appendObject(value) {
+    this.template += '%O';
+    this.substrings.push(value);
   }
 
   /**
@@ -208,7 +279,7 @@ export default class LogModel {
    * @returns {Object}
    */
   getOutputData(level) {
-    this.format(level);
+    this.prepareData(level);
 
     return {
       template: this.template,