+++ src/utils/logger/index.js
@@ -1,3 +1,4 @@
 export { default as loggerClearTemplate } from './loggerClearTemplate';
 export { default as Logger } from './Logger';
 export { default as LogModel } from './LogModel';
+export { default as prepareError } from './prepareError';